+++ tests/unittests/image/test_lpips.py
@@ -49,9 +49,12 @@ def _reference_lpips(
 
     ref = LPIPS(net=net_type)
     res = ref(img1, img2, normalize=normalize).detach().cpu().numpy()
+
     if reduction == "mean":
         return res.mean()
-    return res.sum()
+    if reduction == "sum":
+        return res.sum()
+    return res.flatten()  # for reduction == "none" or None
 
 
 @pytest.mark.skipif(not _TORCHVISION_AVAILABLE, reason="test requires that torchvision is installed")
@@ -61,28 +64,32 @@ class TestLPIPS(MetricTester):
     atol: float = 1e-4
 
     @pytest.mark.parametrize("net_type", ["alex", "squeeze"])
+    @pytest.mark.parametrize("reduction", ["mean", "sum", "none"])
     @pytest.mark.parametrize("ddp", [pytest.param(True, marks=pytest.mark.DDP), False])
-    def test_lpips(self, net_type, ddp):
+    def test_lpips(self, net_type, reduction, ddp):
         """Test class implementation of metric."""
+        if ddp and reduction == "none":
+            pytest.skip("DDP does not support reduction='none'")  # order of gathered tensors is not deterministic
         self.run_class_metric_test(
             ddp=ddp,
             preds=_inputs.img1,
             target=_inputs.img2,
             metric_class=LearnedPerceptualImagePatchSimilarity,
-            reference_metric=partial(_reference_lpips, net_type=net_type),
+            reference_metric=partial(_reference_lpips, net_type=net_type, reduction=reduction),
             check_scriptable=False,
             check_state_dict=False,
-            metric_args={"net_type": net_type},
+            metric_args={"net_type": net_type, "reduction": reduction},
         )
 
-    def test_lpips_functional(self):
+    @pytest.mark.parametrize("reduction", ["mean", "sum", "none"])
+    def test_lpips_functional(self, reduction):
         """Test functional implementation of metric."""
         self.run_functional_metric_test(
             preds=_inputs.img1,
             target=_inputs.img2,
             metric_functional=learned_perceptual_image_patch_similarity,
-            reference_metric=partial(_reference_lpips, net_type="alex"),
-            metric_args={"net_type": "alex"},
+            reference_metric=partial(_reference_lpips, net_type="alex", reduction=reduction),
+            metric_args={"net_type": "alex", "reduction": reduction},
         )
 
     def test_lpips_differentiability(self):
@@ -118,7 +125,7 @@ def test_error_on_wrong_init():
         LearnedPerceptualImagePatchSimilarity(net_type="resnet")
 
     with pytest.raises(ValueError, match="Argument `reduction` must be one .*"):
-        LearnedPerceptualImagePatchSimilarity(net_type="squeeze", reduction=None)
+        LearnedPerceptualImagePatchSimilarity(net_type="squeeze", reduction="invalid_option")
 
 
 @pytest.mark.skipif(not _TORCHVISION_AVAILABLE, reason="test requires that torchvision is installed")